Added Test Cases Change Review Score from Double to Integer Added support for list of genres in MovieBuilder Added Action Movie Genre Overloaded onboardingMovie method to support multiple movie genres Added few missing constructors Flipped movie title and user name in add reviews Added a list users method for movie repository Added Unit test cases according to the mentioned doc

+++ src/main/java/com/crejo/moviereviews/service/analysis/AnalysisServiceImpl.java
@@ -31,7 +31,7 @@ public class AnalysisServiceImpl implements AnalysisService {
     public Double averageReviewScoreByYear(Integer year) {
         return movieRepository.listMovies().stream().filter(movie -> movie.getReleaseYear().equals(year))
                 .map(this::sumOfReviews)
-                .mapToDouble(i -> i)
+                .mapToInt(i -> i)
                 .average().getAsDouble();
     }
 
@@ -39,11 +39,11 @@ public class AnalysisServiceImpl implements AnalysisService {
     public Double averageReviewScoreByMovie(String movieTitle) {
         return movieRepository.listMovies().stream().filter(movie -> movie.getMovieTitle().equals(movieTitle))
                 .map(this::sumOfReviews)
-                .mapToDouble(i -> i)
+                .mapToInt(i -> i)
                 .average().getAsDouble();
     }
 
     private Integer sumOfReviews(Movie movie) {
-        return movie.getReviews().stream().map(Review::getScore).mapToInt(Double::intValue).sum();
+        return movie.getReviews().stream().map(Review::getScore).mapToInt(Integer::intValue).sum();
     }
 }
